Combine note and note types repository CIRC-742
